Rename since there's no more IMS but Fcitx Bridge
