+++ src/candidate_service.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use std::sync::Mutex;
 
-use crate::{candidate::Candidate, ims::FcitxSock};
+use crate::{candidate::Candidate, zmq::FcitxSock};
 
 struct State {
     candidates: Vec<Candidate>,
