+++ src/symbol_service.rs
@@ -2,8 +2,8 @@ use std::sync::Arc;
 
 use std::sync::Mutex;
 
-use crate::ims::FcitxSock;
 use crate::keys::FcitxKeySym;
+use crate::zmq::FcitxSock;
 
 pub struct SymbolService {
     pub(crate) zmq: Arc<Mutex<FcitxSock>>,
